FEAT: added support for angular8, and ivy compatible package
