+++ public_api.ts
@@ -1,0 +1,6 @@
+export * from './src/cheatsheet.component';
+export * from './src/hotkey.model';
+export * from './src/hotkey.options';
+export * from './src/hotkeys.directive';
+export * from './src/hotkeys.service';
+export * from './src/hotkey.module';
