+++ src/hotkey.model.ts
@@ -37,12 +37,12 @@ export class Hotkey {
     /**
      * Creates a new Hotkey for Mousetrap binding
      *
-     * @param {string}   combo       mousetrap key binding
-     * @param {string}   description description for the help menu
-     * @param {Function} callback    method to call when key is pressed
-     * @param {string}   action      the type of event to listen for (for mousetrap)
-     * @param {array}    allowIn     an array of tag names to allow this combo in ('INPUT', 'SELECT', and/or 'TEXTAREA')
-     * @param {boolean}  persistent  if true, the binding is preserved upon route changes
+     * @param combo       mousetrap key binding
+     * @param description description for the help menu
+     * @param callback    method to call when key is pressed
+     * @param action      the type of event to listen for (for mousetrap)
+     * @param allowIn     an array of tag names to allow this combo in ('INPUT', 'SELECT', and/or 'TEXTAREA')
+     * @param persistent  if true, the binding is preserved upon route changes
      */
     constructor(public combo: string | string[], public callback: (event: KeyboardEvent, combo: string) => ExtendedKeyboardEvent | boolean,
                 public allowIn?: string[], public description?: string | Function, public action?: string,
